+++ docs/changelog.md
@@ -32,6 +32,7 @@
 * (Typing) send_message can now reply to any RoomMessage, not just RoomMessageText.
 * `niobot.util.help_command.help_command_callback` was removed, in line with deprecation.
 * `start()` will now query `/_matrix/client/versions` to fetch server version metadata. This is currently unused, however will be helpful for feature detection in the future.
+* Fix RuntimeError due to concurrent typing in send_message
 
 ## v1.1.1 (2024-06-26)